exec: [fix] sign extension in HOST_TR_BASE wedges the core after a VM exit; fix #122
Root Cause:
set_vmcs_host_state() reconstructed the TSS base from the GDT descriptor
with (tr_register->base2) << 24. base2 is an 8-bit bitfield that promotes
to int, so when base2 >= 0x80 the shift lands in the sign bit and the
value sign-extends, destroying the upper 24 bits:

  tss=0xfffffe077c259000 -> 0xfffffe077c259000   (base2 = 0x7c, correct)
  tss=0xfffffe07fc259000 -> 0xfffffffffc259000   (base2 = 0xfc, corrupt)

Per SDM 28.5.2 a VM exit loads TR's hidden base straight from
HOST_TR_BASE, and nothing reloads TR afterwards, so the core carries a
garbage TSS base for the rest of the boot. str and the GDT descriptor are
untouched, which makes the corruption architecturally invisible to any
state check. The measurement itself completes, because RSP0 is only
consulted on a CPL3->CPL0 transition and the whole run is in ring 0 with
interrupts disabled. The first interrupt taken after returning to
userspace reads RSP0 from the garbage base and triple-faults the core,
which then stops responding to INTR and NMI with no oops and no console
output.

cpu_entry_area is randomized per CPU per boot since 6.2, so base2 >= 0x80
is roughly a coin flip. That is why this reproduced on 1 in 2-4 fresh
boots, never on rmmod/insmod within a boot, and only on Intel -- vmrun
does not reload the host TR.

Patch: Widen every descriptor field before shifting.

+++ rvzr/executor_km/x86/vmx.c
@@ -521,8 +521,10 @@ static int set_vmcs_host_state(void)
                  :
                  : "memory");
     struct ldttss_desc *tr_register = (struct ldttss_desc *)(gdtr.address + tr);
-    uint64_t tr_base = ((uint64_t)tr_register->base0 | ((tr_register->base1) << 16) |
-                        ((tr_register->base2) << 24) | ((uint64_t)tr_register->base3 << 32));
+    // each field is widened before shifting; as int bit-fields, base2 << 24 would sign-extend
+    uint64_t tr_base =
+        ((uint64_t)tr_register->base0 | ((uint64_t)tr_register->base1 << 16) |
+         ((uint64_t)tr_register->base2 << 24) | ((uint64_t)tr_register->base3 << 32));
 
     // SDM 25.5 Host-State Area
     // - Control registers
